requirements.txt
Pillow
numpy
opencv-python
matplotlib
scipy
triangle